Add files via upload
Added config options to change the batch sizes of copying, pasting and undoing. The lower you make this value the lower the performance impact on the rest of the server will be, and also the speed of the operation. So on servers with players playing on it a lower value would be better.
Undoing has the biggest performance impact, so on a main server keep this value low.